Move JS for filter into a "component"
This is already done for the CSS, but I forgot the JS. Doing this means
we can more easily use this component in another UI.
