+++ app/frontend/packs/application-provider.js
@@ -2,7 +2,7 @@ require.context("govuk-frontend/govuk/assets");
 
 import { initAll as govUKFrontendInitAll } from "govuk-frontend";
 import initWarnOnUnsavedChanges from "./warn-on-unsaved-changes";
-import filter from "./filter";
+import filter from "./components/paginated_filter";
 import "../styles/application-provider.scss";
 
 govUKFrontendInitAll();
