+++ app/frontend/packs/components/paginated_filter.js
